Move DefaultModelBuildingListener into DefaultProjectBuilder
# Conflicts:
#	maven-core/src/main/java/org/apache/maven/project/DefaultProjectBuilder.java
