+++ maven-core/src/main/java/org/apache/maven/project/DefaultProjectBuilder.java
@@ -71,6 +71,7 @@ import org.apache.maven.api.model.Parent;
 import org.apache.maven.api.model.Plugin;
 import org.apache.maven.api.model.Profile;
 import org.apache.maven.api.model.ReportPlugin;
+import org.apache.maven.api.services.BuilderProblem;
 import org.apache.maven.api.services.MavenException;
 import org.apache.maven.api.services.ModelBuilder;
 import org.apache.maven.api.services.ModelBuilderException;
@@ -81,6 +82,7 @@ import org.apache.maven.api.services.ModelResolver;
 import org.apache.maven.api.services.ModelResolverException;
 import org.apache.maven.api.services.ModelSource;
 import org.apache.maven.api.services.Source;
+import org.apache.maven.api.services.model.ModelBuildingEvent;
 import org.apache.maven.api.services.model.ModelBuildingListener;
 import org.apache.maven.api.services.model.ModelProcessor;
 import org.apache.maven.artifact.Artifact;
@@ -96,6 +98,9 @@ import org.apache.maven.model.building.ModelSource2;
 import org.apache.maven.model.building.ModelSource3;
 import org.apache.maven.model.resolution.UnresolvableModelException;
 import org.apache.maven.model.root.RootLocator;
+import org.apache.maven.plugin.PluginManagerException;
+import org.apache.maven.plugin.PluginResolutionException;
+import org.apache.maven.plugin.version.PluginVersionResolutionException;
 import org.apache.maven.repository.internal.ArtifactDescriptorUtils;
 import org.apache.maven.utils.Os;
 import org.codehaus.plexus.interpolation.AbstractValueSource;
@@ -663,8 +668,7 @@ public class DefaultProjectBuilder implements ProjectBuilder {
             project.setRootDirectory(
                     rootLocator.findRoot(pomFile.getParentFile().toPath()));
 
-            DefaultModelBuildingListener listener =
-                    new DefaultModelBuildingListener(project, projectBuildingHelper, request);
+            ModelBuildingListener listener = new DefaultModelBuildingListener(project, projectBuildingHelper, request);
 
             ModelBuilderRequest modelBuildingRequest = getModelBuildingRequest()
                     .source(ModelSource.fromPath(pomFile.toPath()))
@@ -1410,6 +1414,100 @@ public class DefaultProjectBuilder implements ProjectBuilder {
                 return source;
             } catch (UnresolvableModelException e) {
                 throw new ModelResolverException(e.getMessage(), e.getGroupId(), e.getArtifactId(), e.getVersion(), e);
+            }
+        }
+    }
+
+    /**
+     * Processes events from the model builder while building the effective model for a {@link MavenProject} instance.
+     *
+     */
+    public static class DefaultModelBuildingListener implements ModelBuildingListener {
+
+        private final MavenProject project;
+
+        private final ProjectBuildingHelper projectBuildingHelper;
+
+        private final ProjectBuildingRequest projectBuildingRequest;
+
+        private List<ArtifactRepository> remoteRepositories;
+
+        private List<ArtifactRepository> pluginRepositories;
+
+        public DefaultModelBuildingListener(
+                MavenProject project,
+                ProjectBuildingHelper projectBuildingHelper,
+                ProjectBuildingRequest projectBuildingRequest) {
+            this.project = Objects.requireNonNull(project, "project cannot be null");
+            this.projectBuildingHelper =
+                    Objects.requireNonNull(projectBuildingHelper, "projectBuildingHelper cannot be null");
+            this.projectBuildingRequest =
+                    Objects.requireNonNull(projectBuildingRequest, "projectBuildingRequest cannot be null");
+            this.remoteRepositories = projectBuildingRequest.getRemoteRepositories();
+            this.pluginRepositories = projectBuildingRequest.getPluginArtifactRepositories();
+        }
+
+        /**
+         * Gets the project whose model is being built.
+         *
+         * @return The project, never {@code null}.
+         */
+        public MavenProject getProject() {
+            return project;
+        }
+
+        @Override
+        public void buildExtensionsAssembled(ModelBuildingEvent event) {
+            org.apache.maven.model.Model model = new org.apache.maven.model.Model(event.model());
+
+            try {
+                pluginRepositories = projectBuildingHelper.createArtifactRepositories(
+                        model.getPluginRepositories(), pluginRepositories, projectBuildingRequest);
+            } catch (Exception e) {
+                event.problems()
+                        .add(
+                                BuilderProblem.Severity.ERROR,
+                                ModelProblem.Version.BASE,
+                                "Invalid plugin repository: " + e.getMessage(),
+                                e);
+            }
+            project.setPluginArtifactRepositories(pluginRepositories);
+
+            if (event.request().getRequestType() == ModelBuilderRequest.RequestType.BUILD_POM) {
+                try {
+                    ProjectRealmCache.CacheRecord record =
+                            projectBuildingHelper.createProjectRealm(project, model, projectBuildingRequest);
+
+                    project.setClassRealm(record.getRealm());
+                    project.setExtensionDependencyFilter(record.getExtensionArtifactFilter());
+                } catch (PluginResolutionException | PluginManagerException | PluginVersionResolutionException e) {
+                    event.problems()
+                            .add(
+                                    BuilderProblem.Severity.ERROR,
+                                    ModelProblem.Version.BASE,
+                                    "Unresolvable build extension: " + e.getMessage(),
+                                    e);
+                }
+
+                projectBuildingHelper.selectProjectRealm(project);
+            }
+
+            // build the regular repos after extensions are loaded to allow for custom layouts
+            try {
+                remoteRepositories = projectBuildingHelper.createArtifactRepositories(
+                        model.getRepositories(), remoteRepositories, projectBuildingRequest);
+            } catch (Exception e) {
+                event.problems()
+                        .add(
+                                BuilderProblem.Severity.ERROR,
+                                ModelProblem.Version.BASE,
+                                "Invalid artifact repository: " + e.getMessage(),
+                                e);
+            }
+            project.setRemoteArtifactRepositories(remoteRepositories);
+
+            if (model.getDelegate() != event.model()) {
+                event.update().accept(model.getDelegate());
             }
         }
     }
